Rewrite of station status. More config options.
This is a big update since the initial release of the script:
Re-wrote the entire station notification to be able to notify when the
station has turned off.
Added more options to the config file, such as individual notify on/off
and also
the ability to customize individual notification messages.
Added water level notification.
Added more verbose logging.
Lastly, re-organized the functions and main loop code.